Moved the hook to APReplyBox as they're within same context
